pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "carousel"
dynamic = ["version"]
dependencies = [
    "pytest",
    "pillow"
]